pyproject.toml
[tool.poetry]
name = "mtv"
version = "0.1.0"
description = ""
authors = ["Eric Harm <ericharm@gmail.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.3.0"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
flake8 = "^6.0.0"
attrs = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
